ground_system/hal_service/requirements.txt
RPi.GPIO
pyserial
protobuf